Handle build configuration with react-scripts
The initial application was not built with create-react-app. In order to take benefit from the build configuration provided by create-react-app and stay up to date with the react ecosystem best practices, we switch afterwards to a create-react-app like configuration.

Most of the code stay the same, except for the dependencies in package.json, and some small adjustments to the folder structure:
- `index.html` and images stay in the `public` folder;
- the main javascript file is `index.js` and stays at the root of the `src` (instead of the previous `main.js` file in the `src/js` folder).

To start the development server, we now need to use `npm start` in place of the previous `npm run dev`.